Rename the queue test helper module to testing_utilities
Per review feedback, rename _example_queue.py to testing_utilities.py and update
the imports in the queue tests accordingly.
