+++ tests/dandi_compute_code/queue/test_clean_unsubmitted_capsules.py
@@ -3,7 +3,7 @@ import pathlib
 from unittest import mock
 
 import pytest
-from _example_queue import copy_state_file, create_attempt_directory, entry_for
+from testing_utilities import copy_state_file, create_attempt_directory, entry_for
 
 from dandi_compute_code.queue import clean_unsubmitted_capsules
