+++ tests/dandi_compute_code/queue/test_prepare_queue.py
@@ -4,7 +4,7 @@ import pathlib
 from unittest import mock
 
 import pytest
-from _example_queue import copy_state_file
+from testing_utilities import copy_state_file
 
 from dandi_compute_code.queue import prepare_queue
